Create PeakElement_of_an_Array.cpp
This is a program implementation in C++ to find the peak element in an Array. If the array contains more than one peak element then this program will return any of the peak elements present in the Array.
Time Complexity: O(log n), where n is the number of elements in the Array.
Space Complexity: O(1), as no extra space is used.